Move the `get_config_` function out of `riscv_sim.cpp`.
This helps re-use from other files that might need it.
