+++ c_emulator/config_utils.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <vector>
+#include "sail.h"
+
+uint64_t get_config_uint64(std::vector<const char *> keypath);
